+++ src/logging.ts
@@ -1,7 +1,6 @@
-import Logger from 'genesys-cloud-client-logger';
+import Logger, { ILogger } from 'genesys-cloud-client-logger';
 
 import { GenesysCloudWebrtcSdk } from './client';
-import { ILogger } from './types/interfaces';
 
 export function setupLogging (this: GenesysCloudWebrtcSdk, logger?: ILogger) {
   const {
@@ -10,7 +9,8 @@ export function setupLogging (this: GenesysCloudWebrtcSdk, logger?: ILogger) {
     optOutOfTelemetry,
     originAppId,
     originAppName,
-    originAppVersion
+    originAppVersion,
+    logFormatters
   } = this._config;
 
   this.logger = new Logger({
@@ -22,6 +22,7 @@ export function setupLogging (this: GenesysCloudWebrtcSdk, logger?: ILogger) {
     uploadDebounceTime: 1000,
     initializeServerLogging: !(this.isGuest || optOutOfTelemetry),
     logger,
+    formatters: logFormatters,
     /* consumerApp info */
     originAppId,
     originAppName,
@@ -29,9 +30,9 @@ export function setupLogging (this: GenesysCloudWebrtcSdk, logger?: ILogger) {
   });
 
   if (this.isGuest) {
-    this.logger.debug('Guest user. Not logging to server', null, true);
+    this.logger.debug('Guest user. Not logging to server', null, { skipServer: true });
     return;
   }
 
-  this.logger.debug('Authenticated user. Initializing server logging', null, true);
+  this.logger.debug('Authenticated user. Initializing server logging', null, { skipServer: true });
 }